video-moderation-server/requirements.txt

fastapi
uvicorn
pillow
opennsfw-standalone
